Структура данных для дополнительных предложений вынесена в отдельный файл
